+++ src/config/mod.rs
@@ -4,6 +4,8 @@ mod structs;
 mod config;
 #[path = "impls/finalshell.rs"]
 mod finalshell;
+#[path = "impls/mobaxterm.rs"]
+mod mobaxterm;
 
 pub(crate) use config::*;
 pub(crate) use structs::*;